Add files via upload
The program does everything that I want it to, however, I'm having an issue with infinite loops any time an invalid value is entered. Because of this, and because of the fact that I already have a similar project in my portfolio, I opted to make this repository private and invite you as a collaborator. Also, I completely misunderstood the objective of the 5.5 Final Practical Exam. I read the criteria at the top and misinterpreted it to be criteria for this project before scrolling further. That's my mistake. I'll complete that assignment to the best of my ability before midnight.